refactor: deduplicate download_cifar10.py across examples
Move download_cifar10.py from 5 example directories to a shared scripts/
location. Example directories now import from the shared implementation
using dynamic sys.path manipulation.

Files consolidated:
- scripts/download_cifar10.py (single source)
- examples/{alexnet,googlenet,mobilenetv1,resnet18,vgg16}-cifar10/
  download_cifar10.py (wrapper imports)

Benefits:
- Single point of maintenance for CIFAR-10 dataset handling
- Security fixes apply to all examples instantly
- Reduces code duplication and maintenance burden
- Wrapper scripts preserve original usage patterns
